Refactored code to seperate the API endpoints and ensure GET works successfully
